refactor: domain별로 묶기
