+++ src/main/java/byuntil/backend/admin/domain/Admin.java
@@ -1,0 +1,18 @@
+package byuntil.backend.admin.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+public class Admin {
+    @Id
+    @GeneratedValue
+    @Column(name = "ADMIN_ID")
+    private Long id;
+
+    private String loginId;
+
+    private String loginPw;
+}
