--- src/main/java/byuntil/backend/entity/Board.java
@@ -1,15 +1,0 @@
-package byuntil.backend.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-public class Board {
-    @Id@GeneratedValue
-    @Column(name = "BOARD_ID")
-    private Long id;
-
-    private String name;
-}
