--- src/main/java/byuntil/backend/entity/member/Committee.java
@@ -1,21 +1,0 @@
-package byuntil.backend.entity.member;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Committee extends Member {
-    private String position;
-
-    @Builder
-    public Committee(Long id, String name, String major, String email, String image, String position) {
-        super(id, name, major, email, image);
-        this.position = position;
-    }
-}
