+++ src/main/java/byuntil/backend/member/domain/entity/Category.java
@@ -1,4 +1,4 @@
-package byuntil.backend.entity;
+package byuntil.backend.member.domain.entity;
 
 public enum Category {
     //기계학습, 자연언어처리, 컴퓨터비전, 증강현실AI, 의료AI, 에너지AI
