+++ src/main/java/byuntil/backend/member/domain/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package byuntil.backend.repository;
+package byuntil.backend.member.domain.repository;
 
-import byuntil.backend.entity.member.Member;
+import byuntil.backend.member.domain.entity.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
