+++ src/main/java/byuntil/backend/post/domain/Board.java
@@ -1,4 +1,4 @@
-package byuntil.backend.entity;
+package byuntil.backend.post.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -6,12 +6,11 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 
 @Entity
-public class Admin {
-    @Id@GeneratedValue
-    @Column(name = "ADMIN_ID")
+public class Board {
+    @Id
+    @GeneratedValue
+    @Column(name = "BOARD_ID")
     private Long id;
 
-    private String loginId;
-
-    private String loginPw;
+    private String name;
 }
