+++ src/main/java/byuntil/backend/research/domain/Field.java
@@ -1,5 +1,6 @@
-package byuntil.backend.entity;
+package byuntil.backend.research.domain;
 
+import byuntil.backend.member.domain.entity.Category;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
